Move GetPricesPerShare in right class and refactor some methods, upgrade nuget
