+++ Helpers/FinancialIndicatorCalculator.cs
@@ -89,40 +89,6 @@ public static class FinancialIndicatorCalculator
         return double.IsNaN(currentAssets / currentLiabilities) ? 0 : currentAssets / currentLiabilities;
     }
 
-    public static Dictionary<string, double> GetPricesPerShare(List<string?> symbols)
-    {
-        var prices = new Dictionary<string, double>();
-        PythonEngine.Initialize();
-        using (Py.GIL())
-        {
-            dynamic yf = Py.Import("yfinance");
-
-            var historyData = yf.download(symbols, start: "2019-01-02", end: "2019-01-03");
-
-            foreach (var symbol in symbols)
-            {
-                try
-                {
-                    var tickerData = historyData["Open"].get(symbol);
-                    if (tickerData == null) continue;
-
-                    var price = tickerData.loc["2019-01-02"].As<double>();
-
-                    if (!string.IsNullOrEmpty(symbol) && !double.IsNaN(price))
-                    {
-                        prices[symbol] = price;
-                    }
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine($"Failed to process ticker '{symbol}': {ex.Message}");
-                }
-            }
-        }
-
-        return prices;
-    }
-
     public static double? PriceEarningsRatio(double pricePerShare, List<double?> earningsPerShare)
     {
         if (earningsPerShare.Count == 0) return 0;
